+++ VERSION-HISTORY.md
@@ -1,6 +1,7 @@
 ﻿# 0.0.2
 
 - Name and namespace changed `PostgRest.net` -> `PostgRest.Net` to be more in-line with official convention.
+- Upgrade `Npgsql` NuGet from `4.0.9` to `4.0.10`
 
 # 0.0.1